refactor(types): Phase 2 — LocalTodoItem を types/ へ移動、UUID_REGEX 一元化、AnnotationInput を Omit 導出に置換
ST2-1: LocalTodoItem を hooks/useSupabaseTodos.ts から types/index.ts へ移動。
hooks ファイルパスは変更せず後方互換 re-export を維持。
3コンポーネントのインポートを @/hooks/useSupabaseTodos → @/types に更新。

ST2-2: UUID regex を lib/validation.ts に一元化。
ミューテーション(regex 破壊)で calendar-routes 5件 FAIL を確認→revert 後 PASS。

ST2-3: AnnotationInput を Omit<EventAnnotation, "eventKey"> に置換。
手書き複製を除去し、EventAnnotation との単一真実源を確立。

全テスト 387 passed / tsc エラー 0 を維持。
